Enable hardware unique key (HUK)
Use the HUK API to allow the device to generate
a HUK if one does not already exist. The HUK
will allow the device to encrypt data that only it can decrypt.

Turn off CONFIG_DISABLE_FLASH_PATCH for now
because it doesn't work with CONFIG_HW_UNIQUE_KEY_LOAD.

Add automated tests to allow easy way to build
for all platforms this firmware supports.

bug 21177