Change logging dir and number of files
Logs will be stored in their own folder.
This is more efficient for the logging FS
backend when initializing because it has to
read all files in the dir to determine log file names.

Reduce file size and increase number of files.
This will make it quicker to read files over LwM2M.

bug 22513