Increase BLE LwM2M transport MTU
Increase BLE LwM2M buffer size.
This increases the MTU for LwM2M over BLE. This change allows the gateway to send 1024 byte
CoAP block transfers to a device.
This is important for block transfers sent to an end device.
The first block transfer payload from EdgeIQ is 1024 bytes,
this change ensures the gateway can forward all the
bytes to the end device.

bug 24262

Signed-off-by: Ryan Erickson <ryan.erickson@lairdconnect.com>